==> 1.2.Getting_MRT_coords_thru_onemap.ipynb <==

#!pip install pyshp
#!pip install pyproj
#!pip install geopy
#!pip install rtree
#!pip install pygeos
#!pip install folium
#!pip install geopandas
#!pip install shapely

#Global values

#OneMap header
headers = {'Authorization':'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIyZTcyYWQ5MjFiODI0ZTM4MWNhYzMwODc0MTE2YzQ1ZSIsImlzcyI6Imh0dHA6Ly9pbnRlcm5hbC1hbGItb20tcHJkZXppdC1pdC0xMjIzNjk4OTkyLmFwLXNvdXRoZWFzdC0xLmVsYi5hbWF6b25hd3MuY29tL2FwaS92Mi91c2VyL3Bhc3N3b3JkIiwiaWF0IjoxNzE3NTAxODcyLCJleHAiOjE3MTc3NjEwNzIsIm5iZiI6MTcxNzUwMTg3MiwianRpIjoiSHZ3cEExSlVvR3ZLSmpCciIsInVzZXJfaWQiOjM3MTMsImZvcmV2ZXIiOmZhbHNlfQ.lk2uh3oDJ4EuNHxbhI3oi9i2M5gW5Q82_KIFqGJht94'}

#file output path
fp_output='Relevant_datasets\MRT_coords_B.csv'

#import pygeos
from shapely.geometry import Point
import geopandas as gpd
import os
import pandas as pd
import numpy as np
import requests
import shapefile
import pyproj #https://stackoverflow.com/questions/68317672/coordinate-conversion-script-isnt-giving-me-an-accurate-reading-svy21-to-wgs84
import geopy.distance #https://stackoverflow.com/questions/19412462/getting-distance-between-two-points-based-on-latitude-longitude
from zipfile import ZipFile
from io import BytesIO

import sklearn.neighbors
from sklearn.neighbors import KDTree #https://scikit-learn.org/stable/modules/generated/sklearn.neighbors.KDTree.html

#def Solution:https://stackoverflow.com/questions/5710867/downloading-and-unzipping-a-zip-file-without-writing-to-disk
def get_zip(file_url):
    url = requests.get(file_url)
    zipfile = ZipFile(BytesIO(url.content))
    files = [zipfile.open(file_name) for file_name in zipfile.namelist()]
    return files.pop() if len(files) == 1 else files



#Get coordinates from one map using request (for either latitude or longitude) and location
def getcoord(Req,Location):
    URL="https://www.onemap.gov.sg/api/common/elastic/search?searchVal="+Location+"&returnGeom=Y&getAddrDetails=N"
    r=requests.get(URL,headers=headers)
    json_data=r.json()
    #if onemap returns at least 1 searchval,return the latitude/longitude, else return NA
    if json_data['found']!=0:
        dictdata=json_data['results'][0]
        Latitude=dictdata[Req]
        return Latitude
    else:
        return 'NA'

#MRT=MRT.records()

#Request list of MRT stations from LTA,and using MRT station name to extract coords from onemap for coordinate reference systems consistency
MRT =shapefile.Reader('https://datamall.lta.gov.sg/content/dam/datamall/datasets/Geospatial/TrainStation.zip')
MRT=MRT.records()
MRT=pd.DataFrame(list(MRT),columns=['A','B','Type','MRT_Station'])
MRT=MRT.drop(columns=['A','B'])
MRT

#Create empty columns for latitude and  Longitude to be filled
MRT['Lat']=''
MRT['Lon']=''

#Extract Lat and Lon data from onemap
for index,row in MRT.iterrows():
        Location=MRT['MRT_Station'][index]
        MRT['Lat'][index]=getcoord('LATITUDE',Location)
        MRT['Lon'][index]=getcoord('LONGITUDE',Location)

#EDA

#Visual check 
print(MRT.head(10))
print() #blank row for organization

#check for null or 'NA' data ('NA' is a result of the getcoords function when onemap returns nothing from the query)
print('checking for empty rows...')
print('Null MRT STATIONS:', MRT['MRT_Station'].loc[MRT['Lat'].isnull()])
print('Null MRT STATIONS:', MRT['MRT_Station'].loc[MRT['Lat']=='NA'])
print()

#remove null
MRT=MRT.loc[MRT['Lat'].notnull()]
MRT=MRT.loc[MRT['Lon'].notnull()]
#remove NA
MRT=MRT.loc[MRT['Lat']!='NA']
MRT=MRT.loc[MRT['Lon']!='NA']

#check again for null or 'NA' data after removing NA and null
print('checking for empty rows...')
print('Null MRT STATIONS:', MRT['MRT_Station'].loc[MRT['Lat'].isnull()])
print('Null MRT STATIONS:', MRT['MRT_Station'].loc[MRT['Lat']=='NA'])



#Drop datapoints that are not MRT stations (don't contain the word station e.g. DEPOT)
MRT = MRT[MRT['MRT_Station'].str.contains('STATION')]

#Drop duplicates
MRT.drop_duplicates(keep='first', inplace=False)

#reset index to facilitate smoother ML process later
MRT.reset_index(inplace=True,drop=True)

#Export data
MRT.to_csv(fp_output)
