==> 1.3.Balltree.ipynb <==
#!pip install kdtree
#!pip install scikit-learn==1.0.2

#Import
import os
import pandas as pd
import numpy as np
import seaborn as sns
import requests
import urllib
from urllib.request import urlopen, Request
from sklearn.neighbors import NearestNeighbors
from io import BytesIO
from zipfile import ZipFile
import requests
import shapefile
import statsmodels.api as sm
import statsmodels.formula.api as smf
from enum import unique
import pyproj #https://stackoverflow.com/questions/68317672/coordinate-conversion-script-isnt-giving-me-an-accurate-reading-svy21-to-wgs84
import geopy.distance #https://stackoverflow.com/questions/19412462/getting-distance-between-two-points-based-on-latitude-longitude
import sklearn.neighbors
from sklearn.neighbors import BallTree

##Import datasets

#Import resale flat dataset w/ coor
Address=pd.read_csv("Relevant_datasets/Address_Coord.csv",index_col=0)

#Import MRT coords
MRT=pd.read_csv("Relevant_datasets/MRT_coords_B.csv",index_col=0)

#collapse Latitude and Longitude column into one column for both Resale dataset
Address['Coordinates']=Address.apply(lambda x:np.array([x['Lon'],x['Lat']]) ,axis=1)


##global values

#radius of curvature at the equator ≈6378 km
R = 6371008.7714

#credits to :https://www.vibhuagrawal.com/blog/geospatial-nearest-neighbor-search

#Preprocessing

# convert degrees to radians for Address
Address['rLat']= Address.Lat.apply(np.radians)
Address['rLon'] = Address.Lon.apply(np.radians)

# convert degrees to radians for MRT
MRT['rLat']= MRT.Lat.apply(np.radians)
MRT['rLon'] = MRT.Lon.apply(np.radians)

#credits to :https://www.vibhuagrawal.com/blog/geospatial-nearest-neighbor-search

#Balltree
#Balltree is preferred because of the nature of how coordinates work, which makes calculating distance as if coordinates are points on 2D plane a bad idea.
#Balltree allows for haversine distance to be used, which is a better calculation of distance between 2 points
#haversine*r=distance

#Create balltree model, and set haversine distance (important), leaf size>10 would be helpful
Balltree = BallTree(MRT[['rLat','rLon']].values, leaf_size=15, metric='haversine')
def BallTr(POI):
    #Create a numpy array of coordinates for the point of interest and have it be shape(1,2) - used (-1,2) as it is more robust
    POI=np.reshape(np.asarray(POI),(-1,2))
    #generate distance and index 
    dist, ind = Balltree.query(POI, k=1)
    #Get name of MRT station from index
    MRT_STAT=MRT['MRT_Station'][ind[0][0]]
    #return haversine(dist) and MRT station
    return dist, MRT_STAT

#credits to :https://www.vibhuagrawal.com/blog/geospatial-nearest-neighbor-search


#Create empty var: Closest MRT station name
Address['Cl_MRT']=''
#Create empty var: Distance to closest MRT in KM
Address['Distance_MRT_km']=''
for index,row in Address.iterrows():
    #Experimenting with Lat and Lon
    coord=row[['rLat','rLon']].values
    DIST,MRT_Station=BallTr(coord)
    #Distance=haversine*radius
    Distance=DIST[0][0]*R/1000
    Address.loc[index, 'Distance_MRT_km'] = Distance
    Address.loc[index, 'Cl_MRT'] = MRT_Station

#Visual check of a few values 
Address.head(10)

Address.to_csv('Relevant_datasets/1.3_Address_MRT.csv')
