==> 1.4.Resale_Price_Index.ipynb <==
#Import packages
import pandas as pd

RPI=pd.read_excel("Relevant_datasets/Resale_Price_Index.xlsx")

#realised that there is no data for 2024Q2 , will use the average of the previous 3 quarters as approx

#Get each Qtr's % change from prev Quarter
Q4_2023=RPI['% Change from Previous Quarter'].loc[(RPI['Year']==2023) & (RPI['Quarter']=="IV")].values[0]
Q3_2023=RPI['% Change from Previous Quarter'].loc[(RPI['Year']==2023) & (RPI['Quarter']=="III")].values[0]
Q2_2023=RPI['% Change from Previous Quarter'].loc[(RPI['Year']==2023) & (RPI['Quarter']=="II")].values[0]

#Get average % change
AVG_Delta=(Q4_2023+Q3_2023+Q2_2023)/3

#Get projected index value for 2024 Q2
Proj_Val=RPI['Index'].loc[(RPI['Year']==2024) & (RPI['Quarter']=="I")].values[0]
Proj_Val=Proj_Val*(1+AVG_Delta)

#Create a new row and append to dataset
new_row = pd.DataFrame({'Year': [2024], 'Quarter': ['II'], 'Index': Proj_Val})
RPI = RPI.append(new_row, ignore_index=True)

##Normalize to 2024 Q1

#Normalize using 2024Q1 as index base
index_base=RPI['Index'].loc[(RPI['Year']==2024) & (RPI['Quarter']=="II")].values[0]
#divide every year's index base by 2024Q1's index base, multiply by 100
RPI['Index_2024Q1']=(RPI['Index']/index_base)*100
#drop % change from previous quarter
RPI.drop(columns=['% Change from Previous Quarter','Index'],inplace=True)

#rename index_2024Q1 to index
RPI.rename(columns={'Index_2024Q1':'Index'},inplace=True)

#change syntax for Quarter from roman numerals to numbers
RPI['Quarter']=RPI['Quarter'].apply(lambda x: 1 if (x=="I") else x)
RPI['Quarter']=RPI['Quarter'].apply(lambda x: 2 if (x=="II") else x)
RPI['Quarter']=RPI['Quarter'].apply(lambda x: 3 if (x=="III") else x)
RPI['Quarter']=RPI['Quarter'].apply(lambda x: 4 if (x=="IV") else x)

#Visual check
RPI

#Export
RPI=RPI.to_csv('Relevant_datasets/Resale_Price_Index_Processed.csv')
