==> 1.5_Merge_Address.ipynb <==
#Import packages
import pandas as pd

#global values

##Input filepath for the following folders

#Resale flat prices 1990-1999
fp_90_99="Relevant_datasets/ResaleFlatPricesBasedonApprovalDate19901999.csv"
#Resale flat prices 2000-Feb 2012
fp_00_12="Relevant_datasets/ResaleFlatPricesBasedonApprovalDate2000Feb2012.csv"
#Resale flat prices Mar 2012 - Dec 2014
fp_12_14="Relevant_datasets/ResaleFlatPricesBasedonRegistrationDateFromMar2012toDec2014.csv"
#Resale flat prices Jan 2015-Dec 2016
fp_15_16="Relevant_datasets/ResaleFlatPricesBasedonRegistrationDateFromJan2015toDec2016.csv"
#Resale flat prices Jan 2017 onwards
fp_17_onw="Relevant_datasets/ResaleflatpricesbasedonregistrationdatefromJan2017onwards.csv"

#Address,coordinates and closest MRT folder
fp_Address_coord="Relevant_datasets/1.4_Address_MRT.csv"

#Resale price index
RPI_fp='Relevant_datasets/Resale_Price_Index_Processed.csv'

#output file path for output folder
output_fp="Relevant_datasets/RSF90onw_wAddress.csv"

#Import Address and coordinates folder
Address=pd.read_csv(fp_Address_coord,index_col=0)

###1.Append all datasets together###
##Read 

#read dataframe 1990-1999
RSF_90_99=pd.read_csv(fp_90_99)

#read dataframe 2000-2012
RSF_00_12=pd.read_csv(fp_00_12)

#read dataframe 2012-2014
RSF_12_14=pd.read_csv(fp_12_14)

#read dataframe 2015-2016
RSF_15_16=pd.read_csv(fp_15_16)

#read dataframe 2017 onwards
RSF_17_onw=pd.read_csv(fp_17_onw)


##Creating list of dataframes for easy reference

RSF_list=[RSF_90_99,RSF_00_12,RSF_12_14,RSF_15_16,RSF_17_onw]

#append all datasets together
RSF90_onw=RSF_90_99
for i in range(1,5):
    RSF90_onw=pd.concat([RSF90_onw, RSF_list[i]])

#merge coordinates and Resale flat dataset together
RSF90_onw=RSF90_onw.merge(Address,left_on='street_name',right_on='street_names_oldver',how='left')

#Expect for Nile road to not be mappable from Address dataset to Resale dataset. To drop these rows
print('All unique street names with no closest MRT identified: ',RSF90_onw[RSF90_onw['Cl_MRT'].isnull()]['street_name_x'].unique())
print('number of rows with Nile road as street name to drop: ',RSF90_onw[RSF90_onw['Cl_MRT'].isnull()]['street_name_x'].count())
#Drop rows with Nile Road
RSF90_onw=RSF90_onw.loc[RSF90_onw['street_name_x']!='NILE RD']

print()
print('After dropping...All unique street names with no closest MRT identified: ',RSF90_onw[RSF90_onw['Cl_MRT'].isnull()]['street_name_x'].unique())
print('After dropping...number of rows with Nile road as street name to drop: ',RSF90_onw[RSF90_onw['Cl_MRT'].isnull()]['street_name_x'].count())

#Quick check that the dataframe looks proper
RSF90_onw.head()

### To add resale price normalized by year and quarter to make resale prices comparable across time

##Parse Quarter and Year column from 'month' column##

RSF90_onw['month'] = pd.to_datetime(RSF90_onw['month'])
RSF90_onw['Year'] = RSF90_onw['month'].dt.year
RSF90_onw['Quarter'] = RSF90_onw['month'].dt.quarter

#For ease of merging, will create a column of YYYY_Q
RSF90_onw['Yr_Qt']=RSF90_onw.apply(lambda x: str(x['Year'])+"_"+str(x['Quarter']) ,axis=1)


##Import and process Resale Price index##

#import RPI
RPI=pd.read_csv(RPI_fp,index_col=0)

#Realised that Quarter and Years are floats, changing to integer
RPI['Year']=RPI['Year'].astype('str')
RPI['Quarter']=RPI['Quarter'].astype('str')

#For ease of merging, will create a column of YYYY_Q
RPI['Yr_Qt']=RPI.apply(lambda x: str(x['Year'])+"_"+str(x['Quarter']) ,axis=1)


##Merge Resale Price index and Resale Price dataset on year and quarter and create column of real resale flat price##

#Merge on Yr and Qtr
RSF90_onw=RSF90_onw.merge(RPI,how='left',on='Yr_Qt')

#Generate column of real resale price using price index
RSF90_onw['Real_resale_price']=RSF90_onw['resale_price']*100/RSF90_onw['Index']



RSF90_onw

RSF90_onw.columns

#rename Multi-generation to Multi generation to ensure consistency
RSF90_onw['flat_type']=RSF90_onw.apply(lambda x:"MULTI GENERATION" if (x['flat_type']=='MULTI-GENERATION') else x['flat_type'],axis=1)

RSF90_onw.to_csv(output_fp)
