==> 3.A.ipynb <==
Yishun has received a negative reputation as “Crazy Town”, and property prices might have been impacted. Are Yishun flats the cheapest in the country?

1.To create a dummy variable for each town \
2.encode flat type to integers \
3.To calculate flat age
4.conduct regression

###Import packages
import pandas as pd
import requests
import numpy as np
import statsmodels as sm
import statsmodels.api as smf  
from statsmodels.formula.api import ols #formula guide: https://www.statsmodels.org/dev/examples/notebooks/generated/formulas.html

#global values

#resale fp
RSF_fp="Relevant_datasets/RSF90onw_wAddress.csv"



#Import RSF
RSF=pd.read_csv(RSF_fp,index_col=0,low_memory=False)

RSF['flat_type'].unique()
#to assume a linear relationship btw flat type and resale price.
#assume executive>multi-gen>5 room in general in terms of resale price

#Encode flat type to numerical data
RSF['flat_type_n']=0
RSF.loc[RSF['flat_type']=='1 ROOM','flat_type_n']=1
RSF.loc[RSF['flat_type']=='2 ROOM','flat_type_n']=2
RSF.loc[RSF['flat_type']=='3 ROOM','flat_type_n']=3
RSF.loc[RSF['flat_type']=='4 ROOM','flat_type_n']=4
RSF.loc[RSF['flat_type']=='5 ROOM','flat_type_n']=5
RSF.loc[RSF['flat_type']=='MULTI-GENERATION','flat_type_n']=6
RSF.loc[RSF['flat_type']=='MULTI GENERATION','flat_type_n']=6
RSF.loc[RSF['flat_type']=='EXECUTIVE','flat_type_n']=7
#check number of unchanged rows after encoding
print("no. of unchanged rows:",RSF.loc[RSF['flat_type_n']==0]['flat_type_n'].count())

# to automate the writing of the dummy variable portion of the OLS formula to minimize error

#For easy reference down the line (i.e. when writing regression formula), to replace " " with "_" for town names
RSF['town'] = RSF['town'].str.replace(' ',"_")
RSF['town'] = RSF['town'].str.replace('/',"_")

#create a list of unique towns for easy reference (to be created before turning column to dummy variable)
town_list=list(RSF['town'].unique())

#rename each value in list to the name of its corresponding column by adding "town_"
town_list = ['town_' + x for x in town_list]

#create a dummy variable of towns
RSF_dummies = pd.get_dummies(RSF,columns=['town'])

#convert boolean to int for dummy variables to ease regression
for i in town_list:
   RSF_dummies[i] = RSF_dummies[i].astype(int)

## to automate the writing of the dummy variable portion of the OLS formula to minimize error

#remove Yishun from list before adding to formula to prevent multicollinearity (1 town needs to be absent during regression)
town_list.remove('town_YISHUN')

#checking if yishun is still in the list
if 'town_YISHUN' in town_list:
    print("Yishun is in the list")
else:
    print("Yishun is not in the list")

str_town = ' + '.join(town_list)

print(str_town) 

#Create a variable for flat age
RSF_dummies['flat_age']=RSF_dummies['Year_x']-RSF_dummies['lease_commence_date']

form=str('Real_resale_price ~ flat_type_n + floor_area_sqm + flat_age + '+str_town)
OLS3a = ols(formula=form, data=RSF_dummies).fit()

print(OLS3a.summary())

According to the regression if the dummy variable, after controlling for typical influences of real resale prices such as floor sqm and flat type, there are towns ("town_XXX") with statistically significant correlation coefficients that are negative (i.e. coef<0). That would mean that there are towns were the average prices of resale flats are cheaper than Yishun. Given that the regression table shows that there are towns (e.g. woodlands=-6.433e+04***,sembawang=-3.981e+04***) with statistically significant correlation coefficients<0, it would mean that relative to Yishun, they are cheaper on average. Hence there is no evidence from this analysis that Yishun flats are the cheapest in the country.strong multicollinearity or other numerical problems.
