==> 3.C.1.ipynb <==
Q3C
To create a list of Downtown line MRT and map them to a list of all unique addresses from resale flat data, and identify addresses that are in the same town as the Downtown line stage 2 (DT2) MRT stations

1.Generate list of all DT2 MRT stations (generate list from online sources) \
2.use API request to gather coordinates and Town (through reverse-geocoding Onemap) which they reside in. \
3.Import resale data and keep only addresses of town and street name \
4.Merge with DT2 MRT list \
5.Ensure that all DT2 MRT are mapped properly to a town in resale flat data, and rename if it doesn't (remerge if needed) \
6.Export data

#import packages

import pygeos
from shapely.geometry import Point
import geopandas as gpd
import pandas as pd
import numpy as np
import requests
import urllib
from urllib.request import urlopen, Request
from io import BytesIO
from zipfile import ZipFile
import requests
import shapefile
import pyproj #https://stackoverflow.com/questions/68317672/coordinate-conversion-script-isnt-giving-me-an-accurate-reading-svy21-to-wgs84
import geopy.distance #https://stackoverflow.com/questions/19412462/getting-distance-between-two-points-based-on-latitude-longitude


#Global values

#Resale file path
Resale_fp="Relevant_datasets/RSF90onw_wAddress.csv"

#output dataset
Addr_op='Relevant_datasets/3_C_1_Address_DT2.csv'

#output MRT_DT2
MRT_DT2_op='Relevant_datasets/MRT_DT2.csv'

#OneMap header
headers = {'Authorization':'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJmYjY3Y2U3MTU5MDc1ZDE3MTZhNWViYzQwMTBmNjRjZiIsImlzcyI6Imh0dHA6Ly9pbnRlcm5hbC1hbGItb20tcHJkZXppdC1pdC0xMjIzNjk4OTkyLmFwLXNvdXRoZWFzdC0xLmVsYi5hbWF6b25hd3MuY29tL2FwaS92Mi91c2VyL3Bhc3N3b3JkIiwiaWF0IjoxNzE4NTI0NzM1LCJleHAiOjE3MTg3ODM5MzUsIm5iZiI6MTcxODUyNDczNSwianRpIjoiWWcxa21YSDROQkphbnY2MCIsInVzZXJfaWQiOjM4NjQsImZvcmV2ZXIiOmZhbHNlfQ.CanHC7yILNkuKZbxsSa-48iF6euiIFBLeAFUmzCpIkg'}

##Definitions

#Get coordinates from one map using request (for either latitude or longitude) and location
def getcoord(Req,Location):
    URL="https://www.onemap.gov.sg/api/common/elastic/search?searchVal="+Location+"&returnGeom=Y&getAddrDetails=N"
    r=requests.get(URL,headers=headers)
    json_data=r.json()
    #if onemap returns at least 1 searchval,return the latitude/longitude, else return NA
    if json_data['found']!=0:
        dictdata=json_data['results'][0]
        Latitude=dictdata[Req]
        return Latitude
    else:
        return 'NA'

#Retreive town/planning area where point is in 
def PLN_AREA(LAT,LON):
    URL="https://www.onemap.gov.sg/api/public/popapi/getPlanningarea?latitude="+LAT+"&longitude="+LON
    r=requests.get(URL,headers=headers)
    json_data=r.json()
    #if onemap returns at least 1 searchval,return the latitude/longitude, else return NA
    Town_data=json_data[0]['pln_area_n']
    return Town_data

#create a list of all Downtown line stage 2 MRT stations (released 27th December 2015) and get information on MRT stations they belong to

##1.Generate list of all DT2 MRT stations (generate list from online sources)
#according to : https://en.wikipedia.org/wiki/Downtown_MRT_line
MRT_DT2=['Bukit Panjang MRT Station','Cashew MRT Station','Hillview MRT Station','King Albert Park MRT Station','Sixth Avenue MRT Station','Tan Kah Kee MRT Station','Botanic Gardens MRT Station','Stevens MRT Station','Newton MRT Station','Little India MRT Station','Rochor MRT Station']

#turn list into dataframe
MRT_DT2=pd.DataFrame(MRT_DT2,columns=['MRT_Station'])


##2.use API request to gather coordinates and Town (through reverse-geocoding Onemap) which they reside in.
MRT_DT2['Lat']=''
MRT_DT2['Lon']=''
#Get Latitude and Longitude for each street name through API request to OneMap
for index,row in MRT_DT2.iterrows():
        Location=MRT_DT2['MRT_Station'][index]
        MRT_DT2['Lat'][index]=getcoord('LATITUDE',Location)
        MRT_DT2['Lon'][index]=getcoord('LONGITUDE',Location)

#Get the name of the town where the MRT is in
MRT_DT2['town']=MRT_DT2.apply(lambda x:PLN_AREA(x.Lat,x.Lon),axis=1)

#Create dummy variable for downtown stage 2 MRT to facilitate ease of mapping
MRT_DT2['DT2']=1

MRT_DT2

#Create a list of addresses and their towns, to be merged with MRT list later to see if there are MRT stations that didn't map to any town.

#Read resale flat data
Address=pd.read_csv(Resale_fp,low_memory=False)
#Get town and street name col
Address=Address[['town','street_name_x']]
#Drop duplicates and reset index
Address.drop_duplicates(inplace=True)
Address.reset_index(inplace=True,drop=True)

##After merging  Address with DT2, realised that some MRTs didn't map to any town in resale data, hence went online to search for the town which they reside
#altered the town name after an online search of the closest town they belong to.
##Finding closest named plausible addresses based on their town and street name or google research

#Before changing, to create a column with old town name to facilitate easier mapping later.
MRT_DT2['town_oldver']=MRT_DT2['town']

#Botanic Gardens MRT Station's town is changed to: BUKIT TIMAH
MRT_DT2.loc[MRT_DT2['MRT_Station']=='Botanic Gardens MRT Station','town']='BUKIT TIMAH'

#Stevens MRT Station is changed to: BUKIT TIMAH
MRT_DT2.loc[MRT_DT2['MRT_Station']=='Stevens MRT Station','town']='BUKIT TIMAH'

#Newton MRT Station is changed to: KALLANG/WHAMPOA
MRT_DT2.loc[MRT_DT2['MRT_Station']=='Newton MRT Station','town']='KALLANG/WHAMPOA'

#Little India MRT Station is changed to: KALLANG/WHAMPOA
MRT_DT2.loc[MRT_DT2['MRT_Station']=='Little India MRT Station','town']='KALLANG/WHAMPOA'

#Rochor MRT Station is changed to: KALLANG/WHAMPOA
MRT_DT2.loc[MRT_DT2['MRT_Station']=='Rochor MRT Station','town']='KALLANG/WHAMPOA'

#merge
Addr_MRT=Address.merge(MRT_DT2,on='town',how='left')

#sanity check
Addr_MRT.loc[Addr_MRT['MRT_Station'].notnull()].town.unique()

#fillna for DT2 dummy
Addr_MRT.fillna(value={'DT2':0},inplace=True)

#Keep only address and dummy columns
Addr_MRT=Addr_MRT[['town','street_name_x','DT2','town_oldver']]

#sanity check
Addr_MRT

#export
Addr_MRT.to_csv(Addr_op)
